+++ packages/x-components/src/components/scroll/index.ts
@@ -1,3 +1,3 @@
 export { default as BaseScroll } from './base-scroll.vue';
-export { default as ScrollMixin } from './scroll.mixin';
 export * from './scroll.types';
+export * from './use-scroll';